contactapp/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
dropbox==11.36.2
idna==3.5
ply==3.11
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stone==3.3.1
typing_extensions==4.8.0
urllib3==2.1.0
